Fix calls to kdialog so filters work
The syntax used to supply file extension filters to kdialog (KDE
implementation for _Open/SaveFileDialog$ was entirely wrong. This
now matches what is described in the kdialog manual. Note that my
installed version of kdialog have a bug whereby filters without a
filter description are ignored, so that functionality has not been
tested.