Use parent console if availiable
This change allows QB64 programs to send stdout output to their parent
console if started from one. This works regardless of `$CONSOLE` usage,
which is useful for `_Echo` and `_Log*` statements which can always go
to console regardless of `$CONSOLE` settings.

We also allow a msys or cygwin pipe to be used in place of a proper
console, as that's good enough for our basic output from `_Echo` and
`_Log*` statements (and more convient than starting a separate console
window). If we're an actual `$CONSOLE` program than a separate console
will always be started.