Enable Logging and Terminal configuration from IDE
This allows the user to enable logging from within the IDE. When they do
this, environment variables will be set to turn on logging to console
at Information level with a default set of logging scopes. The program
is started with a separate console to allow the logging to be viewed
separate from the program (unless the program is $Console:Only and has
no graphical view).

The terminal configuration is specific for Linux and used for both
logging output and $Console programs. It allows the user to configure a
terminal emulator to be used for running programs, allowing the output
of $Console programs to be viewed from it as well as logging output when
enabled.

When no terminal is configured we have logic that checks a list of
common terminal emulators to see if any of them are availiable, and if
so uses a known good configuration for them. The user can change
this via a GUI dialog.