Replace wmic with powershell equivalent in Windows set-up scripts.
These should be more forward and backward compatible for us.  (I hope.)  Powershell here is basically just doing the same call to the Win32_Processor and getting the exact same information for us as before.  We're just accessing it with another command, rather than the old wmic command which Windows is depreciating and maybe obsoleting soon.